invoice_streamlit_app/requirements.txt
streamlit
qrcode
python-docx
Pillow